[v1.13 - Komodo] Docker compose support with the Stack resource (#24) Co-authored with @karamvirsingh98
* add some network stuff to container summary

* improve settings tables UI

* periphery build supports additional tags

* fix variable container sizing

* alert types newline wrap

* plumbing for Stack resource

* plumbing for Stack resource

* mount stack api

* stack resource sync

* get remote compose file

* support image_name and image_tag

* add server config placeholders. default server config address

* configure image name and image tag

* deployment work with build image_name and image_tag

* stack UI

* fe builds

* configure registry provider and account

* implement periphery stack api

* stack poll interval

* add UI provider management

* deploy stacks

* build push commit hash tag.

* Destroy stack

* update default core port to 9120

* remove git_account alias

* finish stack (and container) api

* frontend builds

* cant cancel server based builds

* fix

* use git pull -f

* 9120

* start UI updates (#15)

* fix  From<Stack> for CloneArgs

* remove unused imports

* UI Updates (#16)

* cleanup dashboard charts for resources

* bring back solid scrollbars

* enable sidebar scrolling

* remove alerts from all resources

* pass jwt secret

* stacks dont delete the target

* parse services from yaml

* stacks deploy

* close

* looking good

* closer

* destroy stack when file missing. onboard stacks

* figure out stack container name matching

* get stack state correct

* work with service views

* UI Updates - Sidebar, Topbar Alerts, and All Resources page (#17)

* move sidebar to use fixed positioning instead of sticky

* add alert details dialog to topbar alerts

* cleanup all resources page layout

* ensure resource links don't propagate clicks

* periphery support passing env with --env-file

* StackServicePage

* default run_directory to ./ for clarify

* add stack webhook listeners

* add default compose name of stack name

* stacks controlled with project name

* migrate to dotenvy

* add stack to dashboard

* remove deploying / destroying stack services

* update config files

* fix getting service logs

* git / docker provider management api

* implement passing git / registry token from db

* rename system user Github to Git Webhook

* seperate deployed and latest services on stack info

* add stack service level operations

* UI Updates - Update Shadcn/UI components, prevent navbar menu layout shift (#20)

* add dashboard pie for resource syncs

* dashboard items same height

* update shadcn components

* ensure centered following sheet update

* cleanup layout, prevent navbar menu layout shifts

* add manual filter, fix toast call

* guard webhooks

* remove deployed_message, latest_message from StackListItemInfo

* stop all containers on server correctly

* support multiple compose files

* cache all containers networks images projects

* remove project missing from db cache

* work on sync deploy stuff

* rework deployment sync deploy to support stacks. they can depend on each other.

* UI Updates - Remove topbar transparency, pretty status badges, tidy resource page layout with a 'back' button (#21)

* remove topbar transparency

* cleanup unused

* responsive dashboard

* better mobile header

* dont need to calc 64px less since header is using position fixed

* add status badge component

* update status badges

* further simplify layout

* allow undefined status as prop

* use new status badges for alerts

* update status badges for all resources

* undo layout change

* tidy up resource page layout, add back button

* no need for button wrapper

* remove unused

* build cancel log

* update ts types

* fix fe type changes

* fe tweaks

* remove on build logs

* core refresh cache immediately on startup

* jwt_ttl

* canonicalize run directory on host

* update canonicalize error message

* core use docker-compose

* fix incorrect project missing, add status string to stack info

* remove entries in "after" that aren't deploying

* fix dockerfiel

* build custom tag postfix

* sync fixes

* ensure UpdateGitProviderAccount doesn't change id

* ensure UpdateDockerRegistryAccount doesn't change id

*  configure providers in the UI

* add // comment support to env, conversions

* add updates for provider deletes

* improve sync pending deploy log

* add more deployment actions

* add backward compat with v1.12 for clone repo

* stack deploy format

* fe

* alert menus clone when click resource link

* rename stacks

* don't close on click

* snake case stack state, in line with deployment state

* sync redeploy stack if newer hash (optional behind resource field 'latest_hash')

* remove nav to tree

* RefreshStack/Sync debug instruments

* improve inline UI docs

* implement resource base_permission backend

* plumbing for Repo build

* build repos

* write env file repos

* add latest hash / message to build info

* add optional hash to update

* keep built_hash updated

* add backend for build / repo latest hash management

* remove unused resources

* clean up repo dirs after cache update

* fix repo info deser error

* add build / repo git status

* fix page layouts

* improve layout responsive

* most config incline docs

* add descriptions for all resource types

* default local auth false

* fix omnibar arrow keys issue

* add compose file to example config

* image registry

* dashboard display no resources messge

* update deps.

* show when no config

* resource sync use config git_provider

* fix networks

* fix deploy error due to after

* update lots of docs

* fix server stat charts not working

* update screenshots

* update changelog

* add a disclaimer

* remove file paths docs stuff

* build repo

* v1.13 - Komodo

* update docs for cli

* fill out the compose example more

---------

+++ bin/core/src/helpers/sync/remote.rs
@@ -2,11 +2,14 @@ use std::fs;
 
 use anyhow::{anyhow, Context};
 use monitor_client::entities::{
-  sync::ResourceSync, to_monitor_name, toml::ResourcesToml,
-  update::Log, CloneArgs, LatestCommit,
+  sync::ResourceSync, toml::ResourcesToml, update::Log, CloneArgs,
 };
 
-use crate::{config::core_config, state::resource_sync_lock_cache};
+use crate::{
+  config::core_config,
+  helpers::{git_token, random_string},
+  state::resource_sync_lock_cache,
+};
 
 pub async fn get_remote_resources(
   sync: &ResourceSync,
@@ -18,30 +21,28 @@ pub async fn get_remote_resources(
   // commit message
   String,
 )> {
-  let name = to_monitor_name(&sync.name);
   let mut clone_args: CloneArgs = sync.into();
 
   let config = core_config();
 
-  let access_token = match (&clone_args.account, &clone_args.provider) {
+  let access_token = match (&clone_args.account, &clone_args.provider)
+  {
     (None, _) => None,
-    (Some(_), None) => return Err(anyhow!("Account is configured, but provider is empty")),
-    (Some(username), Some(provider)) => config
-      .git_providers
-      .iter()
-      .find(|_provider| {
-        &_provider.domain == provider
-      })
-      .and_then(|provider| {
-        clone_args.https = provider.https;
-        provider.accounts.iter().find(|account| &account.username == username).map(|account| &account.token)
-      })
-      .with_context(|| format!("did not find git token for account {username} | provider: {provider}"))?
-      .to_owned()
-      .into(),
+    (Some(_), None) => {
+      return Err(anyhow!(
+        "Account is configured, but provider is empty"
+      ))
+    }
+    (Some(username), Some(provider)) => {
+      git_token(provider, username, |https| clone_args.https = https)
+        .await
+        .with_context(
+          || format!("Failed to get git token in call to db. Stopping run. | {provider} | {username}"),
+        )?
+    }
   };
 
-  fs::create_dir_all(&config.sync_directory)
+  fs::create_dir_all(&config.repo_directory)
     .context("failed to create sync directory")?;
 
   // lock simultaneous access to same directory
@@ -50,18 +51,25 @@ pub async fn get_remote_resources(
     .await;
   let _lock = lock.lock().await;
 
-  let mut logs =
-    git::clone(clone_args, &config.sync_directory, access_token)
-      .await
-      .context("failed to clone resource repo")?;
+  let repo_path = config.repo_directory.join(random_string(10));
+  // This overrides any other method of determining clone path.
+  clone_args.destination = Some(repo_path.display().to_string());
 
-  let repo_dir = config.sync_directory.join(&name);
-  let LatestCommit { hash, message } =
-    git::get_commit_hash_info(&repo_dir)
-      .await
-      .context("failed to get commit hash info")?;
+  let (mut logs, hash, message, _) = git::clone(
+    clone_args,
+    &config.repo_directory,
+    access_token,
+    &[],
+    "",
+    None,
+  )
+  .await
+  .context("failed to clone resource repo")?;
 
-  let repo_path = config.sync_directory.join(&sync.name);
+  let hash = hash.context("failed to get commit hash")?;
+  let message =
+    message.context("failed to get commit hash message")?;
+
   let resource_path = repo_path.join(&sync.config.resource_path);
 
   let res = super::file::read_resources(&resource_path).map(
@@ -71,8 +79,10 @@ pub async fn get_remote_resources(
     },
   );
 
-  if let Err(e) = std::fs::remove_dir_all(&repo_path) {
-    warn!("failed to remove sync repo directory | {e:?}")
+  if repo_path.exists() {
+    if let Err(e) = std::fs::remove_dir_all(&repo_path) {
+      warn!("failed to remove sync repo directory | {e:?}")
+    }
   }
 
   Ok((res, logs, hash, message))
